Dropped support for Django 1.7.
