+++ .travis.yml
@@ -11,9 +11,6 @@ services:
 env:
   - TOXENV=flake8
   - TOXENV=isort
-  - TOXENV=py27-1.7-sqlite
-  - TOXENV=py27-1.7-postgresql
-  - TOXENV=py27-1.7-postgis
   - TOXENV=py27-1.8-sqlite
   - TOXENV=py27-1.8-postgresql
   - TOXENV=py27-1.8-postgis
@@ -23,21 +20,12 @@ env:
   - TOXENV=py27-master-sqlite
   - TOXENV=py27-master-postgresql
   - TOXENV=py27-master-postgis
-  - TOXENV=py32-1.7-sqlite
-  - TOXENV=py32-1.7-postgresql
-  - TOXENV=py32-1.7-postgis
   - TOXENV=py32-1.8-sqlite
   - TOXENV=py32-1.8-postgresql
   - TOXENV=py32-1.8-postgis
-  - TOXENV=py33-1.7-sqlite
-  - TOXENV=py33-1.7-postgresql
-  - TOXENV=py33-1.7-postgis
   - TOXENV=py33-1.8-sqlite
   - TOXENV=py33-1.8-postgresql
   - TOXENV=py33-1.8-postgis
-  - TOXENV=py34-1.7-sqlite
-  - TOXENV=py34-1.7-postgresql
-  - TOXENV=py34-1.7-postgis
   - TOXENV=py34-1.8-sqlite
   - TOXENV=py34-1.8-postgresql
   - TOXENV=py34-1.8-postgis
