+++ mutant/db/models.py
@@ -2,12 +2,12 @@ from __future__ import unicode_literals
 
 from django.core.exceptions import ValidationError
 from django.db import models
-from django.db.migrations.state import ModelState
+from django.db.migrations.state import ModelState, StateApps
 from django.utils.six import string_types
 from django.utils.translation import ugettext_lazy as _
 
 from .. import logger
-from ..compat import get_remote_field_model, StateApps
+from ..compat import get_remote_field_model
 from ..state import handler as state_handler
