+++ tests/test_model_defs.py
@@ -10,7 +10,6 @@ from django.db.utils import IntegrityError
 from django.test.utils import CaptureQueriesContext
 from django.utils.translation import ugettext as _
 
-from mutant.compat import clear_opts_related_cache, get_related_model
 from mutant.contrib.related.models import ForeignKeyDefinition
 from mutant.contrib.text.models import CharFieldDefinition
 from mutant.db.models import MutableModel
@@ -18,7 +17,7 @@ from mutant.models.model import (
     BaseDefinition, ModelDefinition, MutableModelProxy,
     OrderingFieldDefinition, UniqueTogetherDefinition,
 )
-from mutant.utils import remove_from_app_cache
+from mutant.utils import clear_opts_related_cache, remove_from_app_cache
 
 from .models import (
     AbstractConcreteModelSubclass, AbstractModel, Mixin,
@@ -712,7 +711,7 @@ class BaseDefinitionTest(BaseModelDefinitionTestCase):
             base_definition.base = model_class
             base_definition.save()
         self.assertModelTablesColumnDoesntExists(another_model_class, auto_pk_column)
-        self.assertEqual(get_related_model(model_class.anothermodel.related), another_model_class)
+        self.assertEqual(model_class.anothermodel.related.related_model, another_model_class)
         remove_from_app_cache(another_model_class).mark_as_obsolete()
         self.assertFalse(hasattr(model_class, 'anothermodel'))
         another_model = another_model_class.objects.create(f1='Martinal')
