+++ tox.ini
@@ -1,9 +1,9 @@
 [tox]
 args_are_paths = false
 envlist =
-    py27-{1.7,1.8,1.9,master}-{sqlite,postgresql,postgis},
-    py{32,33}-{1.7,1.8}-{sqlite,postgresql,postgis},
-    py34-{1.7,1.8,1.9,master}-{sqlite,postgresql,postgis},
+    py27-{1.8,1.9,master}-{sqlite,postgresql,postgis},
+    py{32,33}-1.8-{sqlite,postgresql,postgis},
+    py34-{1.8,1.9,master}-{sqlite,postgresql,postgis},
     py35-{1.9,master}-{sqlite,postgresql,postgis}
 
 [testenv]
@@ -25,7 +25,6 @@ deps =
     py32: coverage<4.0
     {py27,py33,py34,py35}: coverage
     redis
-    1.7: django>=1.7,<1.8
     1.8: django>=1.8,<1.9
     1.9: django>=1.9,<1.10
     master: https://github.com/django/django/archive/master.tar.gz
